Migrate breadcrumbs creator to TS
